\name{crop.tri.data}
\alias{crop.tri.data}
\docType{data}
\title{Ethnographic data on the proportions of grain to rachis to weed seeds in the differing crop procesing stages of free-threshing cereals.
}
\description{ Ethnographic data collected in Greece, on the proprotions of free-threshing cereal grain, rachis and weed seeds in four stages during crop processing: winnowing by-products, coarse-seiving by-products, fine-seiving by-products and fine seiving product.
}

\format{ A dataset containing 99 observations on 5 variables (NEED TO FIX THIS FILE TO JUST HAVE FIVE)
\describe{
\item{crop#}{The crop sample number
}\item{Group}{The crop processing stage the sample comes from
}\item{grain}{The total free-threshing cereal grains
}\item{rachis}{The total free-threshing rachis
}\item{weeds}{The total weed seeds}
}}

\source{Jones, G. (1990) The application of present-day cereal processing studies to charred archaeobotanical remains. \emph{Circaea} 6(2):91-96
}

\keyword{datasets}
