\name{CropPro-package}
\alias{CropPro}
\docType{package}
\title{CropPro
\packageTitle{CropPro}
}
\description{
\packageDescription{CropPro}
}

\author{Elizabeth Stroud
\packageAuthor{CropPro}

Maintainer: \packageMaintainer{CropPro}
}
\references{
Reference to in prep paper

Jones, G., 1987. A statistical approach to the archaeological identification of crop processing. \emph{Journal of Archaeological Science}, 14(3), pp.311-323.

Jones, G. 1990. The application of present-day cereal processing studies to charred archaeobotanical remains. \emph{Circaea} 6(2):91-96

Charles, M., 1998. Fodder from dung: the recognition and interpretation of dung-derived plant material from archaeological sites, \emph{Environmental Archaeology}, 1:1, 111-122

}

